Yeah! This version is the first running version that is capable to create an abstract syntax tree from a grammar defined and parsed within libphorward.
lr.parse.c now allows for construction of both an AST or an syntax-tree. Terminals that begin with an "@" are threatened as elements in the AST, or productions that have the attribute 'astname' set. The new production constructor pg_production_create_as_node() allows for AST node linking.

Fine! :) More to come... more detailed, more flexible, more clearness.
This makes fun!!